Add blending for drawing sprite and text

+++ HeartBeat/HeartBeat/Timer.cpp
@@ -18,7 +18,7 @@ void Timer::Update()
 	::QueryPerformanceCounter(reinterpret_cast<LARGE_INTEGER*>(&currentCount));
 
 	sDeltaTime = (currentCount - sPrevCount) / static_cast<float>(sFrequency);
-	sFPS = 1.0f / sDeltaTime;
+	sFPS = static_cast<int>(1.0f / sDeltaTime);
 
 	if (sDeltaTime > 0.05f)
 	{
